style(config): move everything related to modify xp role in package config
