+++ commands/config.go
@@ -2,37 +2,20 @@ package commands
 
 import (
 	"fmt"
-	"github.com/anhgelus/gokord"
 	"github.com/anhgelus/gokord/cmd"
 	"github.com/anhgelus/gokord/component"
 	"github.com/anhgelus/gokord/logger"
 	"github.com/anhgelus/les-copaings-bot/config"
 	"github.com/anhgelus/les-copaings-bot/exp"
 	"github.com/bwmarrin/discordgo"
-	"strconv"
 	"strings"
-	"time"
 )
 
 const (
 	ConfigModify                = "config_modify"
-	ConfigModifyXpRole          = "xp_role"
 	ConfigModifyDisChannel      = "disabled_channel"
 	ConfigModifyFallbackChannel = "fallback_channel"
 	ConfigModifyTimeReduce      = "time_reduce"
-	// xp role related
-	XpRoleAdd       = "xp_role_add"
-	XpRoleAddLevel  = "xp_role_add_level"
-	XpRoleAddRole   = "xp_role_add_role"
-	XpRoleDel       = "xp_role_del"
-	XpRoleDelRole   = "xp_role_del_role"
-	XpRoleEdit      = "xp_role_edit"
-	XpRoleEditLevel = "xp_role_edit_level"
-	XpRoleEditRole  = "xp_role_edit_role"
-)
-
-var (
-	configModifyMap = map[string]uint{}
 )
 
 func Config(_ *discordgo.Session, i *discordgo.InteractionCreate, _ cmd.OptionMap, resp *cmd.ResponseBuilder) {
@@ -97,7 +80,7 @@ func Config(_ *discordgo.Session, i *discordgo.InteractionCreate, _ cmd.OptionMa
 	}).SetComponents(component.New().Add(component.NewActionRow().Add(
 		component.NewStringSelect(ConfigModify).SetPlaceholder("Modifier...").
 			AddOption(
-				component.NewSelectOption("Rôles liés à l'XP", ConfigModifyXpRole).
+				component.NewSelectOption("Rôles liés à l'XP", config.ModifyXpRole).
 					SetDescription("Gère les rôles liés à l'XP").
 					SetEmoji(&discordgo.ComponentEmoji{Name: "🏅"}),
 			).
@@ -121,185 +104,6 @@ func Config(_ *discordgo.Session, i *discordgo.InteractionCreate, _ cmd.OptionMa
 	if err != nil {
 		logger.Alert("config/guild.go - Sending config", err.Error())
 	}
-}
-
-func ConfigModifyXPRole(_ *discordgo.Session, _ *discordgo.InteractionCreate, _ discordgo.MessageComponentInteractionData, resp *cmd.ResponseBuilder) {
-	err := resp.IsEphemeral().
-		SetMessage("Action à réaliser").
-		SetComponents(component.New().Add(component.NewActionRow().Add(
-			component.NewStringSelect(ConfigModify).SetPlaceholder("Action").
-				AddOption(
-					component.NewSelectOption("Ajouter", XpRoleAdd).
-						SetDescription("Ajouter un rôle à XP").
-						SetEmoji(&discordgo.ComponentEmoji{Name: "⬆️"}),
-				).
-				AddOption(
-					component.NewSelectOption("Modifier", XpRoleEdit).
-						SetDescription("Modifier un rôle à XP").
-						SetEmoji(&discordgo.ComponentEmoji{Name: "📝"}),
-				).
-				AddOption(
-					component.NewSelectOption("Supprimer", XpRoleDel).
-						SetDescription("Supprimer un rôle à XP").
-						SetEmoji(&discordgo.ComponentEmoji{Name: "❌"}),
-				),
-		))).Send()
-	if err != nil {
-		logger.Alert("config/guild.go - Sending config", err.Error())
-	}
-}
-
-func ConfigXPRoleAddEdit(_ *discordgo.Session, _ *discordgo.InteractionCreate, data discordgo.MessageComponentInteractionData, resp *cmd.ResponseBuilder) {
-	cID := XpRoleAddLevel
-	if data.CustomID == XpRoleEdit {
-		cID = XpRoleEditLevel
-	}
-	err := resp.IsModal().
-		SetTitle("Role").
-		SetComponents(component.New().ForModal().Add(component.NewActionRow().Add(
-			component.NewTextInput(cID, "Niveau", discordgo.TextInputShort).
-				SetPlaceholder("5").
-				IsRequired().
-				SetMinLength(0).
-				SetMaxLength(5),
-		))).
-		Send()
-	if err != nil {
-		logger.Alert("config/guild.go - Sending modal to add/edit", err.Error())
-	}
-}
-
-func ConfigXPRoleAddRole(_ *discordgo.Session, i *discordgo.InteractionCreate, data discordgo.MessageComponentInteractionData, resp *cmd.ResponseBuilder) {
-	cfg := config.GetGuildConfig(i.GuildID)
-	roleId := data.Values[0]
-	for _, r := range cfg.XpRoles {
-		if r.RoleID == roleId {
-			err := resp.SetMessage("Le rôle est déjà présent dans la config").Send()
-			if err != nil {
-				logger.Alert("commands/config.go - Role already in config", err.Error())
-			}
-			return
-		}
-	}
-	cfg.XpRoles = append(cfg.XpRoles, config.XpRole{
-		XP:     configModifyMap[getKeyConfigRole(i)],
-		RoleID: roleId,
-	})
-	err := cfg.Save()
-	if err != nil {
-		logger.Alert(
-			"commands/config.go - Saving config",
-			err.Error(),
-			"guild_id", i.GuildID,
-			"role_id", roleId,
-			"type", "add",
-		)
-	}
-	if err = resp.IsEphemeral().SetMessage("Rôle ajouté.").Send(); err != nil {
-		logger.Alert("commands/config.go - Sending success", err.Error())
-	}
-}
-
-func ConfigXPRoleEditRole(_ *discordgo.Session, i *discordgo.InteractionCreate, data discordgo.MessageComponentInteractionData, resp *cmd.ResponseBuilder) {
-	cfg := config.GetGuildConfig(i.GuildID)
-	roleId := data.Values[0]
-	_, r := cfg.FindXpRole(roleId)
-	if r == nil {
-		err := resp.SetMessage("Le rôle n'a pas été trouvé dans la config.").Send()
-		if err != nil {
-			logger.Alert("commands/config.go - Role not found (edit)", err.Error())
-		}
-		return
-	}
-	r.XP = configModifyMap[getKeyConfigRole(i)]
-	err := gokord.DB.Save(r).Error
-	if err != nil {
-		logger.Alert(
-			"commands/config.go - Saving config",
-			err.Error(),
-			"guild_id", i.GuildID,
-			"role_id", roleId,
-			"type", "edit",
-		)
-	}
-	if err = resp.IsEphemeral().SetMessage("Rôle modifié.").Send(); err != nil {
-		logger.Alert("commands/config.go - Sending success", err.Error())
-	}
-}
-
-func ConfigXPRoleDel(_ *discordgo.Session, _ *discordgo.InteractionCreate, _ discordgo.MessageComponentInteractionData, resp *cmd.ResponseBuilder) {
-	err := resp.IsEphemeral().
-		SetMessage("Rôle à supprimer").
-		SetComponents(component.New().Add(component.NewActionRow().Add(component.NewRoleSelect(XpRoleDelRole)))).
-		Send()
-	if err != nil {
-		logger.Alert("config/guild.go - Sending response to del", err.Error())
-	}
-}
-
-func ConfigXPRoleDelRole(_ *discordgo.Session, i *discordgo.InteractionCreate, data discordgo.MessageComponentInteractionData, resp *cmd.ResponseBuilder) {
-	cfg := config.GetGuildConfig(i.GuildID)
-	roleId := data.Values[0]
-	_, r := cfg.FindXpRole(roleId)
-	if r == nil {
-		err := resp.SetMessage("Le rôle n'a pas été trouvé dans la config.").Send()
-		if err != nil {
-			logger.Alert("commands/config.go - Sending role not found (del)", err.Error())
-		}
-		return
-	}
-	err := gokord.DB.Delete(r).Error
-	if err != nil {
-		logger.Alert(
-			"commands/config.go - Deleting entry",
-			err.Error(),
-			"guild_id", i.GuildID,
-			"role_id", roleId,
-			"type", "del",
-		)
-	}
-	if err = resp.IsEphemeral().SetMessage("Rôle supprimé.").Send(); err != nil {
-		logger.Alert("commands/config.go - Sending success", err.Error())
-	}
-}
-
-func ConfigXPRoleLevel(_ *discordgo.Session, i *discordgo.InteractionCreate, data discordgo.ModalSubmitInteractionData, resp *cmd.ResponseBuilder) {
-	input := data.Components[0].(*discordgo.ActionsRow).Components[0].(*discordgo.TextInput)
-
-	k := getKeyConfigRole(i)
-	in, err := strconv.Atoi(input.Value)
-	if err != nil || in < 0 {
-		if err = resp.IsEphemeral().
-			SetMessage("Impossible de lire le nombre. Il doit s'agit d'un nombre entier positif.").
-			Send(); err != nil {
-			logger.Alert("command/config.go - Sending bad number", err.Error())
-		}
-		return
-	}
-	configModifyMap[k] = uint(in)
-	go func(i *discordgo.InteractionCreate, k string) {
-		time.Sleep(5 * time.Minute)
-		delete(configModifyMap, k)
-	}(i, k)
-
-	cID := XpRoleAddRole
-	resp.SetMessage("Rôle à ajouter")
-	if data.CustomID == XpRoleEditLevel {
-		cID = XpRoleEditLevel
-		resp.SetMessage("Rôle à modifier")
-	}
-
-	err = resp.IsEphemeral().
-		SetMessage("Rôle à supprimer").
-		SetComponents(component.New().Add(component.NewActionRow().Add(component.NewRoleSelect(cID)))).
-		Send()
-	if err != nil {
-		logger.Alert("config/guild.go - Sending response to add/edit", err.Error())
-	}
-}
-
-func getKeyConfigRole(i *discordgo.InteractionCreate) string {
-	return fmt.Sprintf("r:%s:%s", i.GuildID, i.User.ID)
 }
 
 func ConfigChannel(s *discordgo.Session, i *discordgo.InteractionCreate, optMap cmd.OptionMap, resp *cmd.ResponseBuilder) {
