+++ config/guild.go
@@ -15,13 +15,6 @@ type GuildConfig struct {
 	DaysXPRemains    uint `gorm:"default:90"` // 30 * 3 = 90 (three months)
 }
 
-type XpRole struct {
-	ID            uint `gorm:"primarykey"`
-	XP            uint
-	RoleID        string
-	GuildConfigID uint
-}
-
 func GetGuildConfig(guildID string) *GuildConfig {
 	cfg := GuildConfig{GuildID: guildID}
 	if err := cfg.Load(); err != nil {
